Implemented 3954: AJAX session generation / httponly session ID ; also made sessions in URLs configuration

+++ sources/global2.php
@@ -1615,12 +1615,12 @@ function get_cookie_domain() : string
 /**
  * Get the number of days to store our cookies.
  *
- * @return integer The number of days to store our cookies
+ * @return float The number of days to store our cookies
  */
-function get_cookie_days() : int
+function get_cookie_days() : float
 {
     global $SITE_INFO;
-    return array_key_exists('cookie_days', $SITE_INFO) ? intval($SITE_INFO['cookie_days']) : 120;
+    return array_key_exists('cookie_days', $SITE_INFO) ? floatval($SITE_INFO['cookie_days']) : 120.0;
 }
 
 /**
